Add more CPUID methods.
GCC provides the intrinsics __get_cpuid and __get_cpuid_count.
If __get_cpuid_count is not available (it isn't on my machine) then
__get_cpuid will use whatever value was in ecx at the time for the
subleaf parameter.

Because the corrected version is not available on my machine I do not
think it should be the default method. However, it would be nice to
eventually only rely on intrinsics instead of assenbbly. So I have added
support but not made it default yet.

Additionally, I am keeping the regular inline assembly method around as
a reference. I feel slightly uncertain about this as it imposes a
maintenance burden. But I think it is relatively small and only needs to
be checked along with other manual tests on the manual testing cycle.