Upload test suite results to AppVeyor
The tests were previously being run but their results were not being
displayed in AppVeyor. Displaying those results requires saving the
results to a xml file and uploading it to AppVeyor, using a known
format and a per-run destination URL.

This commit adds the capabiliy to display test results in AppVeyor.