Remove GCC support
GCC is maintained by the Free Software Foundation, which has
supported non-inclusive behavior.

This is against our code of conduct and not the direction our
industry should be headed. As a result, max will not support GCC.
And thus will not indirectly support non-inclusive behavior.

This commit removes GCC support.